add colors/palette projexts
