+++ libs/colors/README.md
@@ -1,0 +1,3 @@
+# Colors
+
+Color manipulation
