+++ libs/light/pxt.json
@@ -22,6 +22,7 @@
     "public": true,
     "dependencies": {
         "core": "file:../core",
+        "colors": "file:../colors",
         "jacdac": "file:../jacdac"
     }
 }
