+++ libs/palette/README.md
@@ -1,0 +1,3 @@
+# Palette
+
+Helpers to manipulate palette in games.
